+++ CHANGES.md
@@ -10,6 +10,7 @@
 * added plugin specific CSS to load from plugin
 * removed `extract()`
 * added `register_activation_hook` to create Voting/Non-Voting roles
+* add 'can_vote' to admin user.
 
 #### 1.9.0